refactor: added information to ReadMe and refactor MPPI stage
